backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
drf-spectacular==0.29.0
python-dotenv==1.2.1
psycopg[binary]==3.3.3
pymysql==1.1.1
openpyxl==3.1.5
stripe==8.7.0
google-auth==2.28.2
